Add uCxAtUtilReplaceChar()
This can be useful for example if you  need to run uCxAtUtilParseParamsF() twice.
uCxAtUtilParseParamsF() will replace each ',' character with a null-terminator,
hence if you execute this function for the same string twice it will not be able
to parse the params on the second call.

With the this new function you can revert the null-termination like this:
  char myParams[] = "123,abc,456";
  size_t orgLen = strlen(myParams);
  uCxAtUtilParseParamsF(myParams, "dsd", ...);
  uCxAtUtilReplaceChar(myParams, orgLen, 0, ',');
  uCxAtUtilParseParamsF(myParams, "dsd", ...);