Remove compiler warning when U_CX_PORT_PRINTF is defined as nothing.
When using uxclient with ubxlib we define U_CX_AT_CONFIG_FILE in order to use our own configuration file. When ubxlib is configured to have no logging at compile-time, so U_CX_PORT_PRINTF is defined to be empty, we get a compiler warning from uCxLogPrintTime() because the stack variables in that function are no longer used. This PR makes the whole of the body of uCxLogPrintTime() conditional on U_CX_PORT_PRINTF being defined in order to remove the warning.